fix(typescript-build): wrong path for dependency
By using the Nuxt's resolver, it was resolving to whatever version of the package `fork-ts-checker-webpack-plugin` npm or yarn would find. That happened because `@nuxt/core` (the package where the Nuxt's resolver is located) doesn't have `fork-ts-checker-webpack-plugin` as a declared dependency. The version installed by `@nuxt/typescript-build` wouldn't always be used if for instance another version of the dependency is also installed on the project.

This fixes #145.